refact : OrderLineItem Quantity 사용하도록 변경

+++ src/main/java/kitchenpos/dto/response/OrderLineItemViewResponse.java
@@ -13,7 +13,7 @@ public class OrderLineItemViewResponse {
                 orderLineItem.getSeq(),
                 orderLineItem.getOrder().getId(),
                 orderLineItem.getMenu().getId(),
-                orderLineItem.getQuantity()
+                orderLineItem.getQuantity().toLong()
         );
     }
